+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#1758](https://github.com/influxdb/influxdb/pull/1758): Add Graphite Integration Test.
 - [#1929](https://github.com/influxdb/influxdb/pull/1929): Default Retention Policy incorrectly auto created.
 - [#1930](https://github.com/influxdb/influxdb/pull/1930): Auto create database for graphite if not specified.
+- [#1908](https://github.com/influxdb/influxdb/pull/1908): Cosmetic CLI output fixes.
 - [#1931](https://github.com/influxdb/influxdb/pull/1931): Add default column to SHOW RETENTION POLICIES.
 
 ### Features